% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{AIDA_QCREF}
\alias{AIDA_QCREF}
\title{Test data for QC of reference genes.}
\format{
\subsection{\code{AIDA_QCREF}}{

\describe{
A dataframe (tibble) with the following columns:
\item{Gene}{Gene name}
\item{Mean_AR_target}{Average ratio of expression from Xi across other candidate reference genes when the SNPs on the gene were used as references}
\item{Mean_AR_reference}{Average ratio of expression from Xi for the gene when SNPs on other candidate reference genes were used as references}
...
}
}
}
\source{
Asian Immune Diversity Network: \url{https://chanzuckerberg.com/science/programs-resources/single-cell-biology/ancestry-networks/immune-cell-atlas-of-asian-populations/}
}
\usage{
AIDA_QCREF
}
\description{
This dataframe (tibble) was generated for the quality control (QC) of reference genes using the AIDA dataset. It was processed using the \code{run_RefGeneQC} function.
}
\keyword{datasets}
