% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{Gene_QC}
\alias{Gene_QC}
\title{QC to remove potential escapee genes from reference gene set}
\usage{
Gene_QC(QCREF, THR = 0.05)
}
\arguments{
\item{QCREF}{Dataframe generated for QC of the reference genes (please also see the \code{run_RefGeneQC} function). QCREF should have the following three columns:
\itemize{
\item Gene
\item Mean_AR_target
\item Mean_AR_reference
}}

\item{THR}{Threshold for the ratio of the expression from Xi.
The default is 0.05, but it should be adjusted according to your dataset.}
}
\description{
QC to remove potential escapee genes from reference gene set
}
