% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{make_per_SNP_sample_df}
\alias{make_per_SNP_sample_df}
\title{Generate a per-SNP summary of the number of samples (for internal use only)}
\usage{
make_per_SNP_sample_df(QCed_df)
}
\arguments{
\item{QCed_df}{Processed scASE data frame generated by the make_QCed_df function}
}
\description{
Generate a per-SNP summary of the number of samples (for internal use only)
}
