% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{make_pseudobulk_refSNP}
\alias{make_pseudobulk_refSNP}
\title{Create pseudobulk data for reference SNPs (for internal use only)}
\usage{
make_pseudobulk_refSNP(DF, REFSNP)
}
\arguments{
\item{DF}{Individual-level scASE data frame}

\item{REFSNP}{ID of the reference SNP}
}
\description{
Create pseudobulk data for reference SNPs (for internal use only)
}
