% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{make_QCed_df}
\alias{make_QCed_df}
\title{Quality Control of the input data frame (for internal use only)}
\usage{
make_QCed_df(DATA, XCI_ref, SNP_DETECTION_DP = 30, SNP_DETECTION_MAF = 0.1)
}
\arguments{
\item{DATA}{Input scASE data frame}

\item{XCI_ref}{XCI status reference data}

\item{SNP_DETECTION_DP}{Threshold for read depth}

\item{SNP_DETECTION_MAF}{Threshold for minor allele frequency}
}
\description{
Quality Control of the input data frame (for internal use only)
}
