% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{multiome_Annotation}
\alias{multiome_Annotation}
\title{Test data for QC of reference genes.}
\format{
\subsection{\code{multiome_Annotation}}{

\describe{
A dataframe (tibble) with the following columns:
\item{cell_barcode}{Cell barcode}
\item{Annotation}{Annotation of the cell}
...
}
}
}
\source{
\url{https://www.10xgenomics.com/resources/datasets/pbmc-from-a-healthy-donor-granulocytes-removed-through-cell-sorting-10-k-1-standard-2-0-0}
}
\usage{
multiome_Annotation
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A dataframe (tibble) for the annotation of the cells in the 10X multiome dataset. Annotation was performed with reference-based mapping with Azimuth <https://azimuth.hubmapconsortium.org/>.
}\if{html}{\out{</div>}}
}
\keyword{datasets}
