% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{multiome_ATAC_ASE_data}
\alias{multiome_ATAC_ASE_data}
\title{Test single-cell allele-specific chromatin accessibility data for scLinaX-multi.}
\format{
\subsection{\code{multiome_ATAC_ASE_data}}{

A dataframe (tibble) containing single-cell allele-specific chromatin accessibility data for 10X multiome data
\describe{
\item{SNP_ID}{SNP identifier}
\item{POS}{Genomic position of the SNP (GRCh38)}
\item{REF,ALT,OTH}{Allelic expression of the reference/alterntive/other allele}
\item{Sample_ID}{Sample_ID}
...
}
}
}
\source{
\url{https://www.10xgenomics.com/resources/datasets/pbmc-from-a-healthy-donor-granulocytes-removed-through-cell-sorting-10-k-1-standard-2-0-0}
}
\usage{
multiome_ATAC_ASE_data
}
\description{
Test single-cell allele-specific chromatin accessibility data for scLinaX-multi. This data was generated from 10X multiome data using cellsnp-lite and annovar.
}
\keyword{datasets}
