% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{multiome_peak_data}
\alias{multiome_peak_data}
\title{Test peak data for scLinaX-multi.}
\format{
\subsection{\code{multiome_peak_data}}{

A dataframe (tibble) with the following columns:
\item{start}{Genomic position of the start of the peak}
\item{end}{Genomic position of the end of the peak}
\item{Peak_name}{Name of the peak}
\item{Gene}{Nearest gene to the peak}
...
}
}
}
\source{
10x Genomics: \url{https://www.10xgenomics.com/resources/datasets/pbmc-from-a-healthy-donor-granulocytes-removed-through-cell-sorting-10-k-1-standard-2-0-0}
}
\usage{
multiome_peak_data
}
\description{

}
\keyword{datasets}
