% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{run_clustering}
\alias{run_clustering}
\title{Generate pseudobulk profiles for individuals and calculate correlations (for internal use only)}
\usage{
run_clustering(
  per_SNP_sample,
  QCed_df,
  Sample,
  valid_ref_genes,
  HE_allele_cell_number_THR = 50,
  QC_total_allele_THR = 10,
  REMOVE_ESCAPE = TRUE,
  PVAL_THR = 0.01,
  RHO_THR = 0.5
)
}
\arguments{
\item{per_SNP_sample}{Output from \code{make_per_SNP_sample_df} function}

\item{QCed_df}{\verb{Output from make_QCed_df} function}

\item{Sample}{Sample ID}

\item{valid_ref_genes}{Output from \code{Gene_QC} function}

\item{HE_allele_cell_number_THR}{Threshold for the number of cells expressing reference SNPs. Candidate reference SNPs expressed in at least "HE_allele_cell_number_THR" cells are used for the analysis. Default: 50.}

\item{QC_total_allele_THR}{Threshold for the total allele count (depth) of the SNP used for calculating the ratio of expression from Xi.
Note that this count is calculated with cells successfully assigned to the group based on the inactivated X chromosome.
This filter is applied in the final step of scLinax and differs from "SNP_DETECTION_DP". Default: 10.}

\item{REMOVE_ESCAPE}{Boolean:
\itemize{
\item TRUE (default): Remove ASE profiles of SNPs on escapee genes when calculating correlations between pseudobulk ASE profiles.
\item FALSE: Include ASE profiles of SNPs on escapee genes when calculating correlations between pseudobulk ASE profiles.
}}

\item{PVAL_THR}{Threshold for P-values and absolute correlation coefficients in the correlation analysis of pseudobulk profiles generated for reference SNPs.}

\item{RHO_THR}{Threshold for absolute correlation coefficients in the correlation analysis of pseudobulk profiles generated for reference SNPs.}
}
\description{
Generate pseudobulk profiles for individuals and calculate correlations (for internal use only)
}
