% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{run_grouping}
\alias{run_grouping}
\title{Define groups for SNP alleles (for internal use only)}
\usage{
run_grouping(CLST1, Merge_list, QCed_df, Sample)
}
\arguments{
\item{CLST1}{Output from the clustering function}

\item{Merge_list}{Output from the clustering function}

\item{QCed_df}{\verb{Output from make_QCed_df} function}

\item{Sample}{Sample ID}
}
\description{
Define groups for SNP alleles (for internal use only)
}
