% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefGeneQC.R
\name{run_RefGeneQC}
\alias{run_RefGeneQC}
\title{Run Gene Quality Control (QC) function}
\usage{
run_RefGeneQC(
  ASE_df,
  XCI_ref,
  SNP_DETECTION_DP = 30,
  SNP_DETECTION_MAF = 0.1,
  SAMPLE_NUM_THR = 3,
  HE_allele_cell_number_THR = 50,
  QC_total_allele_THR = 10
)
}
\arguments{
\item{ASE_df}{A dataframe (tibble) containing single-cell allele-specific expression (scASE) data for all samples. This dataframe should have the following columns:
\itemize{
\item SNP_ID: SNP identifier
\item POS: Genomic position of the SNP (GRCh38)
\item REF: Reference allele of the SNP {A,T,G,C}
\item ALT: Alternative allele of the SNP {A,T,G,C}
\item cell_barcode: Cell barcode
\item REFcount: Allelic expression of the reference allele
\item ALTcount: Allelic expression of the alternative allele
\item OTHcount: Allelic expression of the other allele
\item Sample_ID: Sample ID
\item Gene: Gene annotated to the SNP
}}

\item{XCI_ref}{A dataframe (tibble) containing X chromosome inactivation status. This dataframe should have the following two columns:
\itemize{
\item Gene: Gene name
\item XCI_status: XCI status {escape, variable, inactive}
}}

\item{SNP_DETECTION_DP}{Threshold for the total allele count (depth) of the SNP in the scASE data. SNP–Sample pairs with a total allele count of at least "SNP_DETECTION_DP" are used for the analysis. Default: 30.}

\item{SNP_DETECTION_MAF}{Threshold for the minor allele count of the scASE data. SNP–Sample pairs with a minor allele ratio between "SNP_DETECTION_MAF" and "1 - SNP_DETECTION_MAF" are used for the analysis. Default: 0.1.}

\item{SAMPLE_NUM_THR}{Threshold for the sample size used in the calculation of the ratio of expression from Xi.
Genes evaluated in at least \code{SAMPLE_NUM_THR} samples are used for the calculation of the ratio of expression from Xi. Default: 3.}

\item{HE_allele_cell_number_THR}{Threshold for the number of cells expressing reference SNPs. Candidate reference SNPs expressed in at least "HE_allele_cell_number_THR" cells are used for the analysis. Default: 50.}

\item{QC_total_allele_THR}{Threshold for the total allele count (depth) of the SNP used for calculating the ratio of expression from Xi.
Note that this count is calculated with cells successfully assigned to the group based on the inactivated X chromosome.
This filter is applied in the final step of scLinax and differs from "SNP_DETECTION_DP". Default: 10.}
}
\value{
A dataframe (tibble) with the following columns:
\itemize{
\item Gene: Gene name
\item Mean_AR_target, SD_AR_target: Mean and standard deviation of the ratio of expression from Xi across other candidate reference genes when the SNPs on the gene were used as references
\item Mean_AR_reference, SD_AR_reference: Mean and standard deviation of the ratio of expression from Xi for the gene when SNPs on other candidate reference genes were used as references
\item Mean_Total_allele_target, SD_Total_allele_target: Mean and standard deviation of the total allele count across data points when calculating the ARs defined above for \code{target}.
\item Mean_Total_allele_reference, SD_Total_allele_reference: Mean and standard deviation of the total allele count across data points when calculating the ARs defined above for \code{reference}.
\item Sample_N_target: Number of samples calculating the ARs defined above for \code{target}.
\item Sample_N_reference: Number of samples calculating the ARs defined above for \code{reference}.
\item Count_target: Number of data points calculating the ARs defined above for \code{target}.
\item Count_reference: Number of data points calculating the ARs defined above for \code{reference}.
}
}
\description{
Run Gene Quality Control (QC) function
}
