% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX_multi.R
\name{run_scLinaX_multi}
\alias{run_scLinaX_multi}
\title{Run scLinaX-multi using the scATAC data and output from scLinaX}
\usage{
run_scLinaX_multi(
  scLinaX_summary,
  scLinaX_obj,
  ATAC_ASE_data,
  peak_bed,
  SNP_DETECTION_DP = 30,
  SNP_DETECTION_MAF = 0.1
)
}
\arguments{
\item{scLinaX_summary}{Output from the summarize_scLinaX function run with Annotaion=NULL option}

\item{scLinaX_obj}{A list object generated by the \code{run_scLinaX} function.}

\item{ATAC_ASE_data}{A dataframe (tibble) containing single-cell chromatin accessibility data. This dataframe should include the following columns:
\itemize{
\item SNP_ID: SNP identifier
\item POS: Genomic position of the SNP (GRCh38)
\item REF: Reference allele of the SNP {A,T,G,C}
\item ALT: Alternative allele of the SNP {A,T,G,C}
\item cell_barcode: Cell barcode
\item REFcount: Allelic read count of the reference allele
\item ALTcount: Allelic read count of the alternative allele
\item OTHcount: Allelic read count of the other allele
\item Sample_ID: Sample ID
}}

\item{peak_bed}{A dataframe (tibble) containing single-cell chromatin accessibility data. This dataframe should include the following columns:
\itemize{
\item start: Genomic position of the start of the peak (1-base)
\item end: Genomic position of the end of the peak (1-base)
\item Peak_name: Name of the peak
\item Gene: Nearest gene to the peak
}}

\item{SNP_DETECTION_DP}{Threshold for the total allele count (depth) of the SNP in the scATAC data. SNP–Sample pairs with a total allele count of at least "SNP_DETECTION_DP" are used for the analysis. Default: 30.}

\item{SNP_DETECTION_MAF}{Threshold for the minor allele count of the scASE data. SNP–Sample pairs with a minor allele ratio between "SNP_DETECTION_MAF" and "1 - SNP_DETECTION_MAF" are used for the analysis. Default: 0.1.}
}
\description{
Run scLinaX-multi using the scATAC data and output from scLinaX
}
