% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefGeneQC.R
\name{summarize_QC_gene_table}
\alias{summarize_QC_gene_table}
\title{Summarize the dataframe generated for QC}
\usage{
summarize_QC_gene_table(Ref_Gene_QC, SAMPLE_NUM_THR = 3)
}
\arguments{
\item{Ref_Gene_QC}{Output from the make_QC_gene_table function.}

\item{SAMPLE_NUM_THR}{Threshold for the sample size used in the calculation of the ratio of expression from Xi.
Genes evaluated in at least \code{SAMPLE_NUM_THR} samples are used for the calculation of the ratio of expression from Xi. Default: 3.}
}
\description{
Summarize the dataframe generated for QC
}
