% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX_multi.R
\name{summarize_scLinaX_multi}
\alias{summarize_scLinaX_multi}
\title{Summarize scLinaX-multi results}
\usage{
summarize_scLinaX_multi(phased_df, QC_total_allele_THR = 10, Annotation = NULL)
}
\arguments{
\item{phased_df}{A dataframe generated by the \code{run_scLinaX_multi} function.}

\item{QC_total_allele_THR}{Threshold for the total allele count (depth) of the SNP used for calculating the ratio of expression from Xi. Note that this count is calculated with cells successfully assigned to the group based on the inactivated X chromosome. This filter is applied in the final step of scLinax and differs from "SNP_DETECTION_DP". Default: 10.}

\item{Annotation}{A dataframe (tibble) for the annotation of the cells. This dataframe should have the following columns:
\itemize{
\item cell_barcode
\item Annotation
}}
}
\description{
Summarize scLinaX-multi results
}
