% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_description.R
\docType{data}
\name{XCI_ref}
\alias{XCI_ref}
\title{A dataframe (tibble) containing X chromosome inactivation (XCI) status.}
\format{
\subsection{\code{XCI_ref}}{

A dataframe (tibble) with the following columns:
\describe{
\item{Gene}{Gene name}
\item{XCI_status}{XCI status: escape, variable, inactive}
...
}
}
}
\source{
Tukiainen et al., Nature, 2017; \url{https://doi.org/10.1038/nature24265}
}
\usage{
XCI_ref
}
\description{
This dataframe contains XCI status data, originally sourced from Tukiainen et al., Nature, 2017, with additional manual curation.
}
\keyword{datasets}
