% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{run_scLinaX}
\alias{run_scLinaX}
\title{Run scLinaX analysis for all samples}
\usage{
run_scLinaX(
  ASE_df,
  XCI_ref,
  QCREF,
  Inactive_Gene_ratio_THR = 0.05,
  SNP_DETECTION_DP = 30,
  SNP_DETECTION_MAF = 0.1,
  QC_total_allele_THR = 10,
  HE_allele_cell_number_THR = 50,
  REMOVE_ESCAPE = TRUE,
  PVAL_THR = 0.01,
  RHO_THR = 0.5
)
}
\arguments{
\item{ASE_df}{A dataframe (tibble) containing single-cell allele-specific expression (scASE) data for all samples. This dataframe should have the following columns:
\itemize{
\item SNP_ID: SNP identifier
\item POS: Genomic position of the SNP (GRCh38)
\item REF: Reference allele of the SNP {A,T,G,C}
\item ALT: Alternative allele of the SNP {A,T,G,C}
\item cell_barcode: Cell barcode
\item REFcount: Allelic expression of the reference allele
\item ALTcount: Allelic expression of the alternative allele
\item OTHcount: Allelic expression of the other allele
\item Sample_ID: Sample ID
\item Gene: Gene annotated to the SNP
}}

\item{XCI_ref}{A dataframe (tibble) containing X chromosome inactivation status. This dataframe should have the following two columns:
\itemize{
\item Gene: Gene name
\item XCI_status: XCI status {escape, variable, inactive}
}}

\item{QCREF}{A dataframe generated for reference gene QC (please see the \code{run_RefGeneQC} function). QCREF should have the following three columns:
\itemize{
\item Gene
\item Mean_AR_target
\item Mean_AR_reference
}}

\item{Inactive_Gene_ratio_THR}{Threshold for the ratio of inactive genes}

\item{SNP_DETECTION_DP}{Threshold for the total allele count (depth) of the SNP in the scASE data. SNP–Sample pairs with a total allele count of at least "SNP_DETECTION_DP" are used for the analysis. Default: 30.}

\item{SNP_DETECTION_MAF}{Threshold for the minor allele count of the scASE data. SNP–Sample pairs with a minor allele ratio between "SNP_DETECTION_MAF" and "1 - SNP_DETECTION_MAF" are used for the analysis. Default: 0.1.}

\item{QC_total_allele_THR}{Threshold for the total allele count (depth) of the SNP used for calculating the ratio of expression from Xi.
Note that this count is calculated with cells successfully assigned to the group based on the inactivated X chromosome.
This filter is applied in the final step of scLinax and differs from "SNP_DETECTION_DP". Default: 10.}

\item{HE_allele_cell_number_THR}{Threshold for the number of cells expressing reference SNPs. Candidate reference SNPs expressed in at least "HE_allele_cell_number_THR" cells are used for the analysis. Default: 50.}

\item{REMOVE_ESCAPE}{Boolean:
\itemize{
\item TRUE (default): Remove ASE profiles of SNPs on escapee genes when calculating correlations between pseudobulk ASE profiles.
\item FALSE: Include ASE profiles of SNPs on escapee genes when calculating correlations between pseudobulk ASE profiles.
}}

\item{PVAL_THR}{Threshold for P-values and absolute correlation coefficients in the correlation analysis of pseudobulk profiles generated for reference SNPs.}

\item{RHO_THR}{Threshold for absolute correlation coefficients in the correlation analysis of pseudobulk profiles generated for reference SNPs.}
}
\description{
Run scLinaX analysis for all samples
}
