% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX.R
\name{summarize_scLinaX}
\alias{summarize_scLinaX}
\title{Summarize scLinaX results (return per-individual summary)}
\usage{
summarize_scLinaX(scLinaX_obj, QC_total_allele_THR = 10, Annotation = NULL)
}
\arguments{
\item{scLinaX_obj}{A list object generated by the \code{run_scLinaX} function.}

\item{QC_total_allele_THR}{Threshold for the total allele count (depth) of the SNP used for calculating the ratio of expression from Xi. Note that this count is calculated with cells successfully assigned to the group based on the inactivated X chromosome. This filter is applied in the final step of scLinax and differs from "SNP_DETECTION_DP". Default: 10.}

\item{Annotation}{A dataframe (tibble) for the annotation of the cells. This dataframe should have the following columns:
\itemize{
\item cell_barcode
\item Annotation
}}
}
\value{
A dataframe (tibble) containing scLinaX results.

If no Annotation is supplied, this function returns the ratio of expression from the inactivated X chromosome for all cells.
If Annotation is supplied, this function returns the ratio of expression from the inactivated X chromosome for each cell annotation.

The dataframe includes the following columns:
\itemize{
\item Sample_ID: Sample ID
\item SNP_ID: SNP identifier
\item CHR: PAR or nonPAR
\item POS: Genomic position of the SNP (GRCh38)
\item REF: Reference allele of the SNP {A,T,G,C}
\item ALT: Alternative allele of the SNP {A,T,G,C}
\item Gene: Gene name
\item XCI_status: XCI status {escape, variable, inactive, unknown}
\item Gene_class: XCI status combined with CHR information {PAR1, nonPAR_escape, nonPAR_variable, nonPAR_inactive, nonPAR_unknown, PAR2}
\item Used_as_refGene: Whether the Gene was used as a reference gene {Yes, No}
\item Used_as_refSNP: Whether the SNP_ID was used as a reference SNP {Yes, No}
\item Total_A_allele, Total_B_allele: Total allele count of allele A and B
\item Total_allele: Total allele count of the SNP
\item Expressing_cells: Number of cells expressing the SNP
\item minor_allele_ratio: Ratio of the expression from the allele (A, B) with lower expression (=Ratio of the expression from Xi)
\item Reference_Gene, Reference_SNP: A list of genes used as reference genes and SNPs
\item Num_Reference_Gene, Num_Reference_SNP: Number of reference genes and SNPs
\item Reference_Cell_Count: Number of cells showing mono-allelic expression of reference SNPs
\item Num_A_cells, Num_B_cells: Number of cells showing mono-allelic expression of A and B alleles of reference SNPs
\item Num_Fail_cells: Number of cells showing bi-allelic expression of reference SNPs (should be removed from the analysis)
}

If in per-cell annotation mode, it includes these additional columns:
\itemize{
\item Major_allele: The allele with the higher read count {A, B}
\item Annotation: Cell annotation
}
}
\description{
Summarize scLinaX results (return per-individual summary)
}
