% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLinaX_multi.R
\name{summarize_scLinaX_multi}
\alias{summarize_scLinaX_multi}
\title{Summarize scLinaX-multi results}
\usage{
summarize_scLinaX_multi(phased_df, QC_total_allele_THR = 10, Annotation = NULL)
}
\arguments{
\item{phased_df}{A dataframe generated by the \code{run_scLinaX_multi} function.}

\item{QC_total_allele_THR}{Threshold for the total allele count (depth) of the SNP used for calculating the ratio of expression from Xi. Note that this count is calculated with cells successfully assigned to the group based on the inactivated X chromosome. This filter is applied in the final step of scLinax and differs from "SNP_DETECTION_DP". Default: 10.}

\item{Annotation}{A dataframe (tibble) for the annotation of the cells. This dataframe should have the following columns:
\itemize{
\item cell_barcode
\item Annotation
}}
}
\value{
A dataframe (tibble) containing scLinaX-multi results.

If no Annotation is supplied, this function returns the ratio of expression from the inactivated X chromosome for all cells.
If Annotation is supplied, this function returns the ratio of expression from the inactivated X chromosome for each cell annotation.

The dataframe includes the following columns:
\itemize{
\item Sample_ID: Sample ID
\item SNP_ID: SNP identifier
\item CHR: PAR or nonPAR
\item POS: Genomic position of the SNP (GRCh38)
\item REF: Reference allele of the SNP {A,T,G,C}
\item ALT: Alternative allele of the SNP {A,T,G,C}
\item REFcount: Allelic read count of the reference allele
\item ALTcount: Allelic read count of the alternative allele
\item OTHcount: Allelic read count of the other allele
\item Peak_name: Name of the peak
\item Gene: Nearest gene to the peak
\item XCI_status: XCI status {escape, variable, inactive, unknown}
\item Gene_class: XCI status combined with CHR information {PAR1, nonPAR_escape, nonPAR_variable, nonPAR_inactive, nonPAR_unknown, PAR2}
\item A_allele_count, B_allele_count: Total allele count of allele A and B
\item Total_allele_count: Total allele count
\item minor_allele_ratio: Ratio of the accessible chromatin derived from the allele (A, B) with lower expression (=Ratio of the accessible chromatin derived from Xi)
\item Major_allele: The allele with the higher read count {A, B}
}

If in per-cell annotation mode, it includes the additional column:
\itemize{
\item Annotation: Cell annotation
}
}
\description{
Summarize scLinaX-multi results
}
